virificação de permissão

+++ client/src/views/BaixaPatrimonio.vue
@@ -57,7 +57,7 @@
                 <v-dialog v-model="dialog" max-width="1000px">
 
                     <template v-slot:activator="{ on }">
-                        <v-btn color="primary" class="mb-2"  v-on="on">Novo</v-btn>
+                        <v-btn color="primary" class="mb-2"  @click="novo">Novo</v-btn>
                     </template>
 
                     <v-card>
@@ -218,6 +218,27 @@ export default {
         this.initialize()
     },
     methods: {
+        novo() {
+            if(localStorage.getItem("usuario")) {
+                this.axios.get(process.env.VUE_APP_URL_API + '/permissao/' + localStorage.getItem("usuario")).then(response => {
+                    if(response.data) {
+                        if(response.data.patrimonioDescarregar) {
+                            this.dialog = true;
+                        }
+                        if(!response.data.patrimonioDescarregar) {
+                            this.snackbar = true;
+                            this.color = 'error';
+                            this.textoSnackbar = "Você não tem permissão para baixar patrimônios!";
+                        }
+                    }
+                    if(!response.data){
+                        this.snackbar = true;
+                        this.color = 'error';
+                        this.textoSnackbar = "Tente novamente ocorreu um erro!";
+                    }
+                });    
+            }
+        },
         initialize () {
             this.axios.get(process.env.VUE_APP_URL_API + '/descargaPatrimonio').then(response => {
                 this.desserts = response.data;
@@ -239,26 +260,61 @@ export default {
                 this.patrimonios = response.data;
             });
         },
-
         editItem (item) {
-            this.editedIndex = this.desserts.indexOf(item);
-            this.editedItem = Object.assign({}, item);
-            this.dialog = true;
+            if(localStorage.getItem("usuario")) {
+                this.axios.get(process.env.VUE_APP_URL_API + '/permissao/' + localStorage.getItem("usuario")).then(response => {
+                    if(response.data) {
+                        if(response.data.patrimonioDescarregar) {
+                            this.editedIndex = this.desserts.indexOf(item);
+                            this.editedItem = Object.assign({}, item);
+                            this.dialog = true;
+                        }
+                        if(!response.data.patrimonioDescarregar) {
+                            this.snackbar = true;
+                            this.color = 'error';
+                            this.textoSnackbar = "Você não tem permissão para editar baixas de patrimônios!";
+                        }
+                    }
+                    if(!response.data){
+                        this.snackbar = true;
+                        this.color = 'error';
+                        this.textoSnackbar = "Tente novamente ocorreu um erro!";
+                    }
+                });    
+            }
         },
-
         deleteItem (item) {
-            this.axios.delete(process.env.VUE_APP_URL_API + '/descargaPatrimonio/' + item.id + "/delete").then(response => {
-                if(response.data){
-                    this.snackbar = true;
-                    this.color = 'success';
-                    this.textoSnackbar = "Baixa apagada com sucesso!";
-                    this.initialize();
-                }else {
-                    this.snackbar = true;
-                    this.color = 'error';
-                    this.textoSnackbar = "Ocorreu um erro ao tentar apagar o registro!";
-                }
-            });
+            if(localStorage.getItem("usuario")) {
+                this.axios.get(process.env.VUE_APP_URL_API + '/permissao/' + localStorage.getItem("usuario")).then(response => {
+                    if(response.data) {
+                        if(response.data.patrimonioDescarregar) {
+                             this.axios.delete(process.env.VUE_APP_URL_API + '/descargaPatrimonio/' + item.id + "/delete").then(response => {
+                                if(response.data){
+                                    this.snackbar = true;
+                                    this.color = 'success';
+                                    this.textoSnackbar = "Baixa apagada com sucesso!";
+                                    this.initialize();
+                                }else {
+                                    this.snackbar = true;
+                                    this.color = 'error';
+                                    this.textoSnackbar = "Ocorreu um erro ao tentar apagar o registro!";
+                                }
+                            });
+                        }
+                        if(!response.data.patrimonioDescarregar) {
+                            this.snackbar = true;
+                            this.color = 'error';
+                            this.textoSnackbar = "Você não tem permissão para apagar baixas de patrimônios!";
+                        }
+                    }
+                    if(!response.data){
+                        this.snackbar = true;
+                        this.color = 'error';
+                        this.textoSnackbar = "Tente novamente ocorreu um erro!";
+                    }
+                });    
+            }
+           
         },
 
         close () {
